+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - Enable search transactions by note, address, or txid
 - Move "Export" (export transactions) to account info page
 - ios: fix Pocket user verification button
+- Change block explorer to mempool.space
 
 ## v4.48.6
 - Android: restore support for Android 6 and Android 5